[ADD]routerとroutesのファイルを分類
