+++ plain-go-api/app/routes.go
@@ -1,10 +1,6 @@
 package main
 
-import (
-	"net/http"
-
-	"github.com/gorilla/mux"
-)
+import "net/http"
 
 // routeに対する構造を定義
 type Route struct {
@@ -15,20 +11,6 @@ type Route struct {
 }
 
 type Routes []Route
-
-func NewRouter() *mux.Router {
-
-	router := mux.NewRouter().StrictSlash(true)
-	for _, route := range routes {
-		router.
-			Methods(route.Method).
-			Path(route.Pattern).
-			Name(route.Name).
-			Handler(route.HandlerFunc)
-	}
-
-	return router
-}
 
 // TODO: routerとrouteのファイル分離を行う
 var routes = Routes{
